Klassen an entsprechende Stelle in nach Gradle Architektur verschoben Formatierung auf Repo-Standard geändert, entsprechend IntelliJ defaults
