+++ build.gradle
@@ -3,6 +3,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
+        classpath group: 'de.dynamicfiles.projects.gradle.plugins', name: 'javafx-gradle-plugin', version: '8.8.2'
         classpath 'org.junit.platform:junit-platform-gradle-plugin:1.1.0'
     }
 }
@@ -14,6 +15,8 @@ apply plugin: 'org.junit.platform.gradle.plugin'
 repositories {
     mavenCentral()
 }
+
+apply plugin: 'javafx-gradle-plugin'
 
 dependencies {
     implementation 'com.google.code.gson:gson:2.8.4'
@@ -29,6 +32,11 @@ dependencies {
     testRuntimeOnly(
             'org.junit.jupiter:junit-jupiter-engine:5.1.0'
     )
+}
+
+jfx {
+    mainClass = 'MainController'
+    vendor = '!bigText'
 }
 
 test {
